* Fixed guild and guild member exp truncation issue (bugreport:4130, since r14242). - Fixed guild member position, hair, hair color, gender, class and level could potentially become corrupted due to shorts being read as ints (partially since r2986).
git-svn-id: https://svn.code.sf.net/p/rathena/svn/branches/renewal@14487 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/char/int_guild.h
@@ -14,7 +14,7 @@ int inter_guild_parse_frommap(int fd);
 struct guild *inter_guild_search(int guild_id);
 int inter_guild_mapif_init(int fd);
 int inter_guild_leave(int guild_id,int account_id,int char_id);
-int inter_guild_sex_changed(int guild_id,int account_id,int char_id, int gender);
+int inter_guild_sex_changed(int guild_id,int account_id,int char_id, short gender);
 
 extern char guild_txt[1024];
 extern char castle_txt[1024];
